Wrapped management API methods
Horizon uses module methods for most of API calls. This is
not optimal because then there is no suitable place for business
logic and there is quite a mess now.

This patch moves all module methods into classes, so now
you can call them as class methods. This approach
is cleaner and makes it easier to use api methods (not passing of
request and id attrs to instance methods), also these classes
can be simply extended with other methods.

Change-Id: Ie6a8a4ec528a65e1e714a515afaca6dd6c03c0eb