Added alerts
Alerts data are now mocked in local db. There are precreated 3 alerts
for rack and node - it expects that there will exist rack with id 1 and
nodes with ids 1 and 2. This is for demo purpose.

Currently nodes alerts are not displayed because nodes-related API logic
is not there yet (we return empty array of nodes for each rack).

Change-Id: Ic8a16db2f13909dceb700f2c99269960253da88b