Integrate python-tuskarclient
This patch integrates python-tuskarclient.  To install, run:

pip install -r tools/pip-requires

The patch also updates the ResourceClass and Rack CRUD operations
to use tuskarclient.  'flavors' and 'racks' methods are renamed
to 'list_flavors' and 'list_racks' to prevent collision, as
'racks' is now an attribute of a resource class.  This is
also consistent with Horizon nomenclature.

Associations are *not* updated; so a resource class cannot
associate itself with racks or flavors.  The body of those
methods are simply commented out to allow the UI to 'work'
with no errors.

API tests associated with the above associations fail;
all other tests should pass.

Change-Id: I53dcbb79ad02e32b9bf9d4ed9c61a8dd13e3d0e0