Explicitly convert max_vms to int on resource class creation
Since the resource class creation form doesn't allow editing of the
max_vms fields of the flavours, an attempt to create a class with any
flavours already included will fail, because max_vms will be an empty
string in such cases. I added a check for the empty string and
explicitly convert the value to integer.

Change-Id: If57d01822a85abb1c496bb705fc28644aecbe94a
Closes-Bug: #1220240