Use FormsetStep for selecting flavors in the resource class workflow.
Instead of using data tables, use a django formset for displaying the
list of available flavor templates. There is a new step, FormsetStep
added for the purpose. Additional code for handling validation is added
to the step's action, and code for rendering the formset in a way
similar to the data table is added to the step's template.

Currently the code only allows selecting the templates and setting their
max_vms, but it is possible to modify this solution to allow editing the
class templates in place and even adding new ones directly from the
resource class form.

Change-Id: If57d01822a85abb1c496bb705fc28644aecbe94a
Closes-Bug: #1220240