Create/edit flavors directly in the resource class form.
The formset in the resource class create/edit form now starts empty,
and allows the user to create or edit the flavors of that class
directly. Flavor templates are no longer used.

The formset starts with a single empty row at the end, for adding a new
flavor, and a "Delete" column with checkboxes for deleting existing
flavors. If javascript is available, there is also a link for adding
more empty rows.

TODO: If javascript is available, replace the checkboxes with a cross
icon, for deleting the rows instantly. Also, remove the empty row at
the end, as the users can always add it with the link.

Change-Id: I1ff4c9f90657c62e7c5f45a4b7874524b9df204c
Implements: blueprint remove-flavor-templates