Remove as much of custom DataTable code as possible
Since tuskar-ui now uses the FormsetStep for the flavors in the
resource class create/edit dialog, a lot of the custom DataTable
code is no longer needed, so I removed it.

There is still a custom Row class needed for pre-selecting the
checkboxes in the Racks table in the resource class create/edit.

Change-Id: I51647309ddccafba1d8e07ad12586e6bf2d1badd