Separated out node and baremetal nodes in api
The tuskar representation of a Node contains two ids;
'id' is for the tuskar representation, and 'nova_baremetal_node_id'
is for the baremetal node.  The UI confuses the two,
receiving the tuskar id and using it to query nova baremetal.

This patch explicitly separates the tuskar Node representation
from the BaremetalNode representation.  To make sure
there is no confusion in the testing data, the test data
Nodes and BaremetalNodes use independent ids

Change-Id: I671f731c769d8a0af5fb6b02017bf5f414e9c42d