Corrected Node and BaremetalNode methods and attributes
This patch corrects the Node and BaremetalNode methods and
attributes.  It also updates the UI and tests.

Doing this simultaneously is a necessity, as the key
relationship change here is that a Node has a one-to-one
relationship with a BaremetalNode, through which the UI
fetches many attributes for display.  Moving the attributes
without implementing the relationship (and vice versa) causes
the UI and tests unusable.

Change-Id: I090fbc1305a0a61e003f5b3b2aea44c9bb156116
Closes-Bug: #1236192