Renames get_* methods for consistency
get_<object> methods are there because the object in
question is also an attribute.  This commit removes the
attribute from the class' list of attributes and renames
the get_<object> method.

The only time the attribute is needed is when the code
wants the id of <object> without fetching the entire object.
Those <object>_id methods reach directly into the _apiresource
in order to get that id.

horizon does something similar in its nova api code.

Change-Id: Idbfee52ce6163c61acbf0102e01766bb59983f1e