Rename "node" variables and methods to tuskar_node or baremetal_node
Since we have so many different types representing node data in some
way, it has became hard to read the code when it's not certain what
kind of node we are dealing with at a particular point. Renaming the
variables to clearly indicate the node type should help with it.

This patch only does the changes that don't require changing anything
in tuskar or tuskarclient. There are still a few places left that
require such changes, but that requires some synchronization.

Change-Id: If5fae7b60318ec91fd288da5a42bdefb93fce911
Partial-bug: #1244202