Fix tuskar/baremetal node confusion on node deleting
The code that deletes a node from a rack had the tuskar and baremetal
nodes confused, which became visible after renaming all the variables.
This patch fixes that confusion, fixes a bug that was introduced during
the rename that made it impossible to delete the last node from a rack,
and moves the node removal code into the api.py.

Change-Id: I2d85b4826a74579221a1e149c3f43f9e60a93b87