Use baremetal node id for node detail page
Currently, the node detail page is keyed off the
tuskar node id.  However, a tuskar node merely maps
a rack to a baremetal node, and does not persist if
the rack is deleted (while the baremetal node does).
This means that the detail page for an unracked node
will error out.

This patch updates the node detail page (and relevant
links) to use the baremetal node id.

Change-Id: I69a32226b3848b45b1a58417ea4e209c28c3e43a
Closes-Bug: #1252813