Revert "Support building wheels (PEP-427)"
This reverts commit f4f51b2421031564e6d8136badb686a79b26535a.

The patch being reverted should not have been merged, because
setting universal to True is used to claim in-place python 2/3
compatibility in Wheel packages.

We don't support python 3 yet, so we must not claim the opposite
here (as this would harm end users as soon as we start uploading
Wheel packages to PyPi).

Please see the following links for details:
 - http://pythonwheels.com/
 - http://lists.openstack.org/pipermail/openstack-dev/2013-November/020059.html

Change-Id: I81bedf16744bc66701a6882dcd32ca3814262bf7