Group related panels in packages
Make the file structure of the source reflect the logical
structure of the application (and its menus) by grouping
related panels into packages. This gives us a place to put
common code shared by those panels, and lets us find the
source code for a specific panel based on where it is in
the menu.

Because Horizon was not initially designed for autodiscovery
of panels that are not immediate children of the dashboard,
the URLs of those panels will have the form of
/infrastructure/resources.archived/. That can be amended by
a simple patch to Horizon later on.

Change-Id: I8df74ac8c8ffb913b9e653f049f64d725c4e32d3