Make the panels nested
In order to appeal to Ruby On Rails aesthetics more,
where flat is not necessarily better than nested,
I moved our panels into "deployment" and "resources"
directories. This makes the URLs for them look slightly
strange, like /infrastructure/resources.archived/, but
I think this is the best we can do to force such a
structure.

Alternatively we could just leave it flat, obeyong the
Zen of Python, which says:

Flat is better than nested.

Change-Id: I8df74ac8c8ffb913b9e653f049f64d725c4e32d3