Group related panels in packages (resources)
This is a part of a larger patch, split to make it less
confusing. This patch deals with moving the files into
the "resources/" direcotry.

Make the file structure of the source reflect the logical
structure of the application (and its menus) by grouping
related panels into packages. This gives us a place to put
common code shared by those panels, and lets us find the
source code for a specific panel based on where it is in
the menu.

Because Horizon was not initially designed for autodiscovery
of panels that are not immediate children of the dashboard,
the URLs of those panels will have the form of
/infrastructure/resources.archived/. That can be amended by
a simple patch to Horizon later on.

Change-Id: I843efc86b582fb7b9e2d046dc83edefbebb5a861