Add mock api calls to the UI
This is an introductory commit that adds expected API
calls to tuskar-ui and has them return test data.  It
is intended as a base to allow further discussion, as
well as allow parallel development on the UI side and
API side.

There are many questions raised by this patch - for
example:

* what are the parameters of the tuskarclient api?
* can Ironic create a Node with a single api call?
* etc

But those fall out of scope of this particular patch,
and will be resolved in later patches.

Change-Id: I1273c8738b760ff36d5ad21e3982beb10457c011