Update power_state method to return Ironic states
Currently, there is a temporary workaround that lets Tuskar
run with Nova baremetal instead of Ironic.  The power_state
method is intended to 'translate' the baremetal task_state to
the Ironic power_state.  This patch ensures that the return
value of the method matches the possible values of the Ironic
node attribute.

Change-Id: Ib43578e5bc5074f88ec348fd2519454704f12839