Add list/created/delete operations for node profiles
Add separate tab with list of all node profiles
and buttons to create/edit/delete them.

Node profiles implementation is based on Nova flavor
management code in Horizon itself.

Note that updating node profiles is not supported yet,
as it causes several issues (indeed, flavors are never
updated, they are recreated with new ID, breaking
the links from overcloud roles to flavors and
confusing the users). Policy on how to update
node profiles is to be created lately.

Change-Id: I4041a7f59ab9a0abdcb66c7ae4d752284b472ad9