Proof of concept of error handling in the API
Currently error handling in Tuskar-UI involves a lot of repeated
code, as the same pattern is used everywhere and the same API calls
tend to be handled in the same way no matter where they appear.

This patch introduces a decorator that can be used to add default
error handling to all API calls that take a request as a parameter.

That default error handling can be disabled or modified using the
special parameters passed to the call, all of which start with
"_error" to avoid conflicts.

OvercloudRole.list is decorated to show how it is supposed to be used
and how it simplifies the code.

Change-Id: Ie441789c471deeb6d64267bf4424f5661ef073af