Use "python -m coverage" rather than use "coverage" explicitly
When coverage module is installed using a package from Debian/Ubuntu,
the command name is "python-coverage", and there is no /usr/bin/coverage
command. However, using "python -m coverage" doesn't work in a Python
2.6 virtual env.

This patch tests for the presence a "coverage" command line utility,
if it isn't there, it will try to search for "python-coverage", and
finally and if there's none, then it uses "python -m coverage".

Change-Id: Ia25a77a0556c87d73a85a2804f00985cde0ee622
Closes-bug: #1292789