Improve deployment page
Use num_nodes attribute of overcloud.counts to determine number of
nodes associated with each role on the overcloud overview page.
If this not available, fall back to the old way of using len(nodes).

Also, improve the deployment template to show accurate information
about number of nodes currently deployed, deploying, waiting for
deployment, and in error state.

Change-Id: I8ce533d6f7229af551be3a42a0adca8c54909e13
Closes-bug: #1288193