Improve error reporting
In case of errors, the qunit testcase extracts the test name,
error message and source of the error from the qunit page
and prints them, so that it's easier to locate the issue.

Also, the selenium.py file got renamed to qunit.py, because
importing the selenium library fails if the importing script
is called selenium.py (python tries to import the library
from the importing script itself).

Change-Id: I040d147ad9c78fe48080b6a01ff1b4d89d6e9c3e