Silence the deprecation warning about with statements
The tests use contextlib.nested for nesting multiple with statements,
but that function generates a warning in python 2.7 telling us, that
there is a better syntax for it. Unfortunately, we need to support
older python versions, so we can't use that new syntax. The warning is
annoying, as it clobbers the output of tests and makes other warnings
harder to see. This patch silences that warning in tests.

Change-Id: I6e49faad4c45c13ee2e0041fe9d31b7fcd3670a6