Update node index view to match Juno wireframes
For now, instead of having a single 'Registered' tab with a graphical/
table switch, we have two separate tabs.  The former is under 'Overview';
the latter under 'Registered'.

The graphical portion of the 'Overview' is left for a future patch.

'Discovered' tab is eliminated pending a better understanding of the
Ironic API that enables it.

Additional styling to come in future patches.

Partial-Implements: blueprint node-index-view-update
Change-Id: Id62493ab7dfb8cee97905c59d457e8b1bdc72ecf