Added Overview panel
The overview panel is a combination of the former overcloud and plan
panels.

The assumption here is that a plan already exists; different view states
will be determined by whether an associated stack exists.

A future patch will have to update api.tuskar.Plan.get_the_plan to create
a plan if none exists.

Change-Id: I37d0943d3dbe0cb2849af82a459bc5b2af17881c