Remove Neutron init from post-deployment init
The post-deployment init always creates a default-net tenant network,
a router and an ext-net external network. While this is good for demo
purposes they are not needed in a production environment for naming
or functional reasons and first thing that an operator does is to delete them.

You might not need the default router. For instance if you want to test vlan
provider networks. But the init creates the ext-net network and default router
though they are not needed.

It generates confusion when using network isolation. You have to specify
floating ip cidr, floating ip start, floating ip end, bm network gateway,
that match the external network settings set up in the network isolation
yaml file in order to get working connectivity. And you may end up deleting
them right after init finishes so you get a bad user experience.

Closes-bug: https://bugzilla.redhat.com/show_bug.cgi?id=1243056
Change-Id: Iddfcc39498ce1ac991b103df0b32b49f074deb34