Rename context_selection to prevent coliding with Horizon
When new templatetag gets added to context_selection.py in Horizon
it is missing when using tuskar-ui because we override the module.
This patch renames it in Tuskar-ui so it won't collide. To make it
work, new templatetag module needs to be loaded in Tuskar-ui base.html

I have created patch in Horizon that refactors iframe_embed_settings
templatetag to separate module as it does not fit into context_selection.
https://review.openstack.org/205122
When it gets merged, we might consider reverting this patch, although I
think it would be better to keep it.

Change-Id: I92e2ee477a44360f68a15599fe73e9361bb7a991