Allow Tenant Switching
This is achieved by removing context_selection templatetag module,
that we used to override horizon behavior

Due to this bug https://bugs.launchpad.net/horizon/+bug/1023502
it is not possible to determine in which tenant the user will get
logged in. So we need to let user change tenant if needed.

Change-Id: Ib7f7076678e6ec51a88dca7bbd9a472adbadaf2a